# bridge/shared/mobile/openapi.yaml
openapi: 3.1.0
info:
  title: OMEGA-ARC Mobile Operator API
  version: 0.4.2
servers:
  - url: /api/mobile/v1
security:
  - bearerAuth: []
paths:
...
  /conversations/active:
    get:
      responses:
        '200': { description: Most recently active runtime conversation }
        '404': { description: No conversation exists }
  /conversations/{conversationId}:
    get:
      parameters:
        - in: path
          name: conversationId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Conversation history }
  /conversations/{conversationId}/messages:
    post:
      parameters:
        - in: path
          name: conversationId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content, client_message_id]
              properties:
                content: { type: string, minLength: 1 }
                client_message_id: { type: string }
      responses:
        '200':
          description: Existing OMEGA-ARC SSE stream (phase, delta, learning, confidence, end, or error payloads)
          content:
            text/event-stream:
              schema: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...